feat: centralizando lógica de eventos capturados em GameEventListener
